+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added implementation for `Brep.from_pipe` in `compas_rhino.geometry.RhinoBrep`.
 * Added implementation for `Brep.from_iges` in `compas_rhino.geometry.RhinoBrep`.
 * Added implementation for `Brep.to_step` in `compas_rhino.geometry.RhinoBrep`.
+* Added implementation for `Brep.to_viewmesh()` in `compas_rhino.geometry.RhinoBrep`.
 
 ### Changed